[MIRROR] [MDB IGNORE] Fixes a runtime from `can_track` on mobs on the syndicate base (#5438)
Original PR: https://github.com/tgstation/tgstation/pull/74847
-----
## About The Pull Request

`user` is not guaranteed to be passed and can be `null`.

This cause a runtime, when there is a mob on the syndicate base and they
are being checked for trackability.

Fixes it with one character `null` check. 

This is safe, because `THING in null` fails. Any mobs on syndie base are
default untrackable if there is no `user`.
(Famous last words)

## Why It's Good For The Game

Runtime

## Changelog

:cl: Melbert
fix: Runtime from tracking mobs on the syndicate base
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>