[MIRROR] [MDB IGNORE] Restores ability for borgs to adjust reflector angles (#5483)
Original PR: https://github.com/tgstation/tgstation/pull/74913
-----
## About The Pull Request
Adds an override for `attack_robot()` on reflectors, which is called
when a borg clicks on an object with no tool selected. Only allows
rotating if the borg is adjacent to the reflector.
## Why It's Good For The Game
This was an undocumented change of #74800, and I assume unintended. In
any case, the inability to adjust reflectors somewhat restricts engineer
borgs' ability to engineer.
## Changelog
:cl:
fix: Restores ability for borgs to adjust reflector angles. This is now
a left-click with nothing selected, similar to humans' left-click with
an empty hand.
/:cl:

---------

Co-authored-by: zxaber <37497534+zxaber@users.noreply.github.com>