[MIRROR] [MDB IGNORE] March into Mapness - Moon Base 19 (#5505)
Original PR: https://github.com/tgstation/tgstation/pull/74067
-----
Hello again gang.

## About The Pull Request

This PR is a remake of #68031. I've been away for... a bit... but March
Mapness reminded me I had this project stalled, and I wanted to get this
one out the door.




![image](https://user-images.githubusercontent.com/41715314/225826044-57f36412-97f7-4aca-b69d-7c4200a8277f.png)
Sneak Previews of one of the reworked areas.

![image](https://user-images.githubusercontent.com/41715314/232185113-1fb4cade-f3b2-497a-9bde-85263286a047.png)
Explore new sights!

![image](https://user-images.githubusercontent.com/41715314/232185282-6f941e2c-e7dd-48ae-9e5a-fd517296be2c.png)
Puzzles!?

![image](https://user-images.githubusercontent.com/41715314/232185338-b3035eb3-96ff-4a3b-9891-f1bc0e6641b3.png)
And more!

Features (taken from the old PR body:
**Lunar Sand!**
Unlike the old yellow colored ocean sand that we had on the outpost, the
base now has lunar-accurate solar-washed gray sand. It behaves in much
the same way as regular basalt or yellow sand, but with more of a rocky
texture and darker hanging shadows over the rocky surface.

**Alien Plants and Weeds!**
Uniquely, the caves and xenoflora of the moon outpost have grown over
the xenomorph weeds, giving a tiny, unique bit of diversity to the usual
xenoflora of weeds, nests, and nodes.

**Misc Objects and Sprites!**
Fluff meteorites have been added as static objects, not spinning and
hurdling at you for a coup de gras, but instead as a fluff object that
has landed and being studied by the moon outpost 19 staff. Additionally,
some new rock sprites were also added to the random rock pool to help
diversify the map pool since the rocky outcrop of the moon features a
lot of... rocks.

**Soul Pass and Var Edits!**
Some meticulous detail was used in overseeing the map and the objects
contained therein. This includes some custom names, descriptions, lots
of unique placement, and better atmospheric storytelling of the tragedy
that befell this research outpost before the xenomorph attack was
underway. This is primarily an expansion of the existing rooms and nooks
and crannies of the old map as well as some of the new rooms added.

Oh yeah new rooms and areas were added!
This map was already a 255x255 tile space, but very, VERY little of the
map was used before it was originally implemented. Now, we've added a
few new rooms and challenges to the map to help in both the flavor of
the map and giving players more to do on the map itself. The main 3
areas added are:

The ATV Shed, used for extended outdoor missions on the moon outpost
away from the research station. Seems like the shed was locked up tight
after the last office party however, and the late RD was the last one to
have the key...
Ground Zero, the original excavation site, where basalt and rocks from
Lavaland broke off and landed on the moon, resulting in an investigation
as these rocks were thousands of years older than modern deposits from
Lavaland. However, it seems that something found it's way back as well.
The Lunar Caves, connecting the northern and southern sites together.
The monsters appear to have set up a forward base here, catching the
remaining stragglers off guard. Will you be the next?
The Old Digshaft, since abandoned, appears to be rattling with life.
With all the other hissing and cracking nearby, surely it can't be worse
than anything you've already fought, right?

### Mapping March

Ckey to receive rewards: arcanemusic

## Why It's Good For The Game

Moon Outpost 19 is a vastly empty map that hasn't gotten much love in a
great while now. This gave me a chance to give this a new coat of paint
between the lunar rocks and sand, rock types, some new aesthetic flora,
and to add some much needed personality to the map as it stands today.
Plus, I've had this sitting on my harddrive for about 8 months now, so
I'd hope I can get this shipworthy.

## Changelog

:cl:
add: A few map specific lunar objects have been added related to Konsu
19 (Moon Outpost 19).
add: New fluff objects were added related to xenoflora and mining.
add: New loot, puzzles, and challenges await you on Konsu 19. Are you
prepared?
/:cl:

---------

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>