[MIRROR] [MDB IGNORE] Shuttle loan fix (#5569)
Original PR: https://github.com/tgstation/tgstation/pull/75020
-----
## About The Pull Request

shuttle_loan_datum.dm had two entries for the department_resupply
variation. One of those is the Syndicate hijacking event, and the other
is the resupply. This fixes that so those two events don't conflict with
each other.

## Why It's Good For The Game

It removes a problem with the code, and makes it so that admins can
force either the syndicate hijacking or the resupply

## Changelog

:cl:
fix: Shuttle Loan event's resupply and hijacking have been separated
into two different datums
/:cl:

---------

Co-authored-by: norsvenska <73006946+norsvenska@users.noreply.github.com>
Co-authored-by: tattle <66640614+dragomagol@users.noreply.github.com>