[MIRROR] [MDB IGNORE] Ectoplasm Anomaly Name and Icon Bugfix (#5646)
Original PR: https://github.com/tgstation/tgstation/pull/75162
-----
## About The Pull Request
Does as the name states. I found it weird that the ectoplasm anomaly
core is not referred to as raw when you have to refine it and instead
used both the normally refined core icon as well as the name not saying
it was a raw anomaly core at all.
## Why It's Good For The Game
Oversights are bad right? This aims to solve one of them.
## Changelog
:cl:
fix: fixed unrefined ectoplasm anomaly cores not appearing as unrefined
/:cl:

---------

Co-authored-by: RealityOverseer <76832653+RealityOverseer@users.noreply.github.com>