Fixes for the all-american diner ruin (#75202)
## About The Pull Request

- Fixes accesses for the all-american diner
- Fixes airlock cycling for the same

## Why It's Good For The Game

Currently the ruin requires the deep storage "bunker access ID" or
similar to open doors. This seems like it's probably an oversight; it's
not exactly a ruin bursting with swag and seems to be more just a
thematic place to hang out (and the access helper is, perhaps,
counterintuitively named?). So, this removes the access requirements
from all doors except the "boss's room", making it a bit more friendly
as a hangout instead of a place where every door must be busted open in
one way or another.

The external-access airlocks were cycling as separate pairs rather than
all together. This has also been fixed.

## Changelog
:cl:
fix: The All-American Diner no longer requires a bunker ID card for most
rooms nor requires accesses to use the jukebox.
fix: The All-American Diner's external airlocks now all work together.
/:cl: