[MIRROR] [MDB IGNORE] Fixes a runtime from foam in a weird loc / improves the stack trace (somewhat) (#5699)
Original PR: https://github.com/tgstation/tgstation/pull/75281
-----
## About The Pull Request

- Fixes a runtime from foam being created inside plasmamen, instead of
at the plasmamen's loc, from the clown plasmaman suit


![image](https://user-images.githubusercontent.com/51863163/236899740-1d5ccca3-43c4-455b-9bb2-3206ce433df8.png)

- Adds a stack trace for mobs having foam created inside of them
- So, in the past foam would runtime if created in a non-turf at all,
mobs or otherwise. I changed it so it creates a more helpful stack
trace.
- While tackling this, however, I realized that we may want to support
foam in non-turf, non-mob locs? Like closets? That is to say, if you
release a foam reaction while stuck in a locker and it floods the
locker, it would stay in the locker until it is opened, and affect other
people in the locker itself? Unsure, so I left it as an option, albeit
it may not function entirely correctly.
- This may also false-positive in some places, such as a mob within mob.
I'm unsure if it should have its own case, or of it should act like
closets are implied to do above? But that may produce some silent
errors. So eh.

## Why It's Good For The Game

Less runtimes, more intended code. 

## Changelog

:cl: Melbert
fix: Fixes a runtime with plasmaman clown suit lube spray
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>