Mortidrobe Restocks are now in Medical Wardrobe Supply Crates (#75299)
I noticed the mortidrobe restock wasn't actually in the medical wardrobe
supply crate, so i added it to that. now coroners can get restocks for
their clothes through cargo. the restock item itself, and the part to
make it work with the vending machine were already done, so it was
likely intended to be obtainable via this crate.
Also fixes a typo of "autospy" to "autopsy" in the autopsy kit crate.