Fixes material sniffers not being able to find materials properly (#75354)
![image](https://github.com/tgstation/tgstation/assets/6209658/d91d07ba-32b3-4580-a22a-616e0aab55e2)

stack/init was adding stacks to the global list if they had LESS THAN 10
units, though the sniffer only looks for things >= 10 units. Flipped
bracket, gg. add() was also using += instead of |=, which resulted in
duplicate entries for the same object in the global list when adding
stuff to a stack.

:cl: ShizCalev
fix: The materials sniffer now actually works!
/:cl: