[MIRROR] [MDB IGNORE] Adds the Syndicate Induction Kit for Nuclear Operatives (#5823)
Original PR: https://github.com/tgstation/tgstation/pull/75055
-----
## About The Pull Request

Adds the Syndicate Induction Kit for Nuclear Operatives.

The kit contains the following

- One Ansem pistol
- One each of 10mm AP/HP magazines
- An energy dagger
- A syndicate bowman headset
- A nuclear PDA and agent card
- A syndicate fabric space suit and helmet, alongside a jet harness
- Combat boots, turtleneck, fingerless gloves
- The Fission Mailed book
- And the most important part, the induction implant.

The induction implant will make anyone who is implanted by it a
Syndicate Operative. This will give them the syndie faction and the nuke
op datum, alongside the usual weird name nukeops get.

The implant will NOT work on anyone who is not an antagonist in some way
or form. Normal crew members do not gain benefits from implanting, and
in fact the implant disintegrating inside gives them toxin.

It costs a total of 10 TC, for what is more or less 25 TC of gear. The
value is totally up to change, it's probably way too low.

Uplink spacesuits are randomized between all their different colored
variants! Visual only.

## Why It's Good For The Game

It's cool and rewards teamwork. Traitors can team up with nukies, and if
they choose to do so they gain some basic gear (a free gun, a space
suit, the dagger) that isn't really even close to the degree of gear
real ops get. As a downside, their name becomes an extremely obvious
nuclear operative name, meaning if they want to be stealthy they either
have to never talk or purchase a voice changer.

Being able to just obtain a whole new member of the team for so little
likely seems bonkers, but my intention is to just encourage something
that's neat and interesting ingame without hopefully being *too*
impactful and strong.

Imagine stumbling upon a heretic and inducting them into the team. Or a
changeling, or a wizard? Running around hypnoflashing people to be nuke
ops and then giving them the kit?? There's a lot of fun moments that
could arise from this, but they're balanced by the fact that, well,
nobody ever saves TC for the station.

Something I can see becoming a problem is EVERY traitor buying syndicate
encryption keys to beg for an induction kit every time war is declared.
I'm not sure if this is going to be a huge issue, but I think I'd rather
just get this ingame and see if that does turn out to be a problem. If
it does, we can just make it so that traitors can't be inducted, as the
syndicate has already hired them for something else and doesn't want to
flare inter-faction tensions.

## Changelog

:cl:
add: Adds the Syndicate Induction Kit for Nuclear Operatives, which lets
them induct any willing antagonist into the nuclear operative team for
10 TC.
add: Uplink spacesuits are randomized between all their different
colored variants! Visual only.
/:cl:

---------

Co-authored-by: carlarctg <53100513+carlarctg@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>