[MIRROR] [MDB IGNORE] Refactored the TTS subsystem to more properly handle message garbling. Added a volume preference for TTS. (#5906)
Original PR: https://github.com/tgstation/tgstation/pull/75559
-----
## About The Pull Request
TTS subsystem now uses datums to handle requests. Also removed the
DM-side caching since it wasn't really all that useful compared to the
complexity it added to the entire system.

Messages will no longer play at the same time the request was received.
This is so that the TTS audio files don't play out of order.
Also adds a preference for players to adjust the volume of TTS sounds.
Additionally adds a debug verb to re-establish connection to the TTS
server.

Removes double processing when calculating the listeners for TTS
messages as well, which may fix some issues where you'd be able to hear
a message that you weren't supposed to hear.
Fixed being able to hear messages whilst unconscious by refactoring the
Hear proc so that it returns a boolean value for whether the message was
understood or not.

## Why It's Good For The Game
Fixes message garbling. Also gives players a way to control the volume
of TTS messages for themselves.

## Changelog
:cl: Watermelon914, Iamgoofball
fix: Fixed message garbling. TTS messages won't play in the wrong order
anymore.
qol: Added a way for players to change the volume of TTS.
/:cl:

---------

Co-authored-by: Watermelon914 <37270891+Watermelon914@users.noreply.github.com>
Co-authored-by: Watermelon914 <3052169-Watermelon914@users.noreply.gitlab.com>
Co-authored-by: Iamgoofball <iamgoofball@gmail.com>
Co-authored-by: Kyle Spier-Swenson <kyleshome@gmail.com>