[MIRROR] [MDB IGNORE] Refactors snipping cuffs into a bespoke cuffsnapping element (#5945)
Original PR: https://github.com/tgstation/tgstation/pull/75432
-----

## About The Pull Request

Refactors snipping cuffs into a bespoke cuffsnapping element, adding
support for delayed cuffsnipping. Adds this element to box cutters!
Effectively speaking everything is the same as usual.

## Why It's Good For The Game

It's cool, it's based and elementized and modularized and not
hardcodeized on the jaws of life anymore. Plus it could be used in the
future for things (it won't)

## Changelog

:cl:
refactor: Refactors snipping cuffs into a bespoke cuffsnapping element,
adding support for delayed cuffsnipping. Adds this element to box
cutters! Effectively speaking everything is the same as usual.
/:cl:

---------

Co-authored-by: carlarctg <53100513+carlarctg@users.noreply.github.com>