[MIRROR] [MDB IGNORE] The Kevin De-cleodening: Custom elevator music (#5948)
Original PR: https://github.com/tgstation/tgstation/pull/75518
-----
## About The Pull Request
I got the idea while riding an elevator the other day to make a remix of
Title2 but Elevator-style. So I downloaded FLStudio, spent a couple of
hours fighting with it, made a song, and opened this PR up. Later on,
cats4gold contacted me saying they had the same exact idea, Title2 and
all, and after I suggested the ide of combining both versions, we get
the status of this PR today!

Here's a video of me riding an elevator (Exciting!):

https://youtu.be/BNQmTe553lU

Also our current looping sound system is very very bad, the problem
people have with hearing elevator music outside the elevator is because
the checks for stuff like volume falloff only happen when an audio
starts playing, so if you stand besides an elevator button and hear the
music, if you go far away you'll still hear it as if you're next to it.
I don't know how to fix this so I didn't touch it, so hopefully someone
else that does, does! :)

Here's a link to the full track:
https://vocaroo.com/1idoRor8G1xY

## Why It's Good For The Game
Way more charming, way more immersive.
## Changelog
:cl: TheSmallBlue and cats4gold
sound: Recent research concluded that the new elevator music was upping
crewmate's morale a bit too much, this has now been corrected via the
replacement of said music.
code: Added a new "in_order" variable to the looping sound datum, for
when you need sounds to play in order
/:cl:

---------

Co-authored-by: TheSmallBlue <ilanmori@hotmail.com>