[MIRROR] [MDB IGNORE] The Canvas UI now shows a grid while the user is painting. (#5965)
Original PR: https://github.com/tgstation/tgstation/pull/75604
-----
## About The Pull Request
A grid is shown while the user is holding a painting tool on an editable
canvas. This is achieved by drawing empty rectangles after the filled
ones are placed whenever the canvas is updated and if requirements are
met.

The result is more or less like this:


![immagine](https://github.com/tgstation/tgstation/assets/42542238/fa8fe495-eac7-4947-84af-cc04e953da4b)

It's by no mean perfect, especially since the strokeStyle is a mere
"#888888" hexcolor, the contrast may get low around colors of similar
lightness. But I don't think I've much the strength to seek out a
fancier solution to a peeve that's hardly a botherance anyway while the
canvas is still mostly white.

I've also had a maintainer suggest me to make a toggle to
disable/enable. I talked a bit how that'd be a bit more complex to
integrate but not much so frankly, although switching hands or putting
away the tool already works for that. That said, I'm willing to do it if
player feedback suggests such toggle should be integrated. EDIT: Fine,
done!

## Why It's Good For The Game
This should help players tell the distance between one point and another
in the canvas without as much guessing and mistakes, whether it's
copyism, cropped images, or original compositions.

## Changelog

:cl:
qol: The Canvas UI now shows a grid while the user is painting.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>